queue updates

+++ database/migrations/0002_01_01_000006_create_buildings_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignId('details_id')
                   ->constrained('building_details')
                   ->onDelete('cascade');
-            $table->decimal('effect_value')->nullable();
+            $table->bigInteger('effect_value')->nullable();
             $table->integer('level')->default(1);
             $table->integer('build_time')->nullable();
             $table->timestamps();
